zephyr: only allow recovery to format fs
move fs_shell to recovery, and add a format command for recovery mode
