+++ app/prj.conf
@@ -21,3 +21,6 @@ CONFIG_LOG=y
 
 # Enable shell
 CONFIG_SHELL=y
+
+# diable ability to format filesystem
+CONFIG_FILE_SYSTEM_MKFS=n
