+++ app/src/coriander/backends/zephyr/zephye_extra_shell.cc
@@ -1,0 +1,10 @@
+/**
+ * @file zephye_extra_shell.cc
+ * @author Savent Gate (savent_gate@outlook.com)
+ * @brief
+ * @date 2023-08-17
+ *
+ * Copyright 2023 savent_gate
+ *
+ */
+#include "../../recovery/src/fs_shell.c"
