+++ recovery/prj.conf
@@ -1,1 +1,7 @@
-# nothing here
+CONFIG_USE_SEGGER_RTT=y
+
+CONFIG_LOG=y
+
+CONFIG_SHELL=y
+CONFIG_SHELL_LOG_BACKEND=y
+CONFIG_SHELL_BACKEND_RTT=y
